moved SofastService to SofastGUI so that it may be used as a comparison
